+++ src/data/mockTrendingBooks.js
@@ -1,0 +1,185 @@
+const trendingBooks = {
+  query: "/trending/daily",
+  works: [
+    {
+      key: "/works/OL17590212W",
+      title: "The Subtle Art of Not Giving a F*ck",
+      edition_count: 36,
+      first_publish_year: 2016,
+      has_fulltext: true,
+      public_scan_b: false,
+      ia: ["subtleartofnotgi0000mans_n2n5", "subtleartofnotgi0000mans_h2r6"],
+      ia_collection_s: "internetarchivebooks;printdisabled",
+      cover_edition_key: "OL26193423M",
+      cover_i: 8231990,
+      language: ["eng", "por", "rus", "spa"],
+      author_key: ["OL2964000A"],
+      author_name: ["Mark Manson"],
+      availability: {
+        status: "private",
+        available_to_browse: false,
+        available_to_borrow: false,
+        available_to_waitlist: false,
+        is_printdisabled: true,
+        is_readable: false,
+        is_lendable: false,
+        is_previewable: true,
+        identifier: "subtleartofnotgi0000mans_n2n5",
+        isbn: "9780062899149",
+        oclc: null,
+        openlibrary_work: "OL17590212W",
+        openlibrary_edition: "OL27351482M",
+        last_loan_date: null,
+        num_waitlist: null,
+        last_waitlist_date: null,
+        is_restricted: true,
+        is_browseable: false,
+        __src__: "core.models.lending.get_availability",
+      },
+    },
+    {
+      key: "/works/OL17930368W",
+      title: "Atomic Habits",
+      edition_count: 20,
+      first_publish_year: 2018,
+      has_fulltext: true,
+      public_scan_b: false,
+      ia: [
+        "atomic-habits-tiny-changes-remarkable-results-by-james-clear-z-lib.org",
+      ],
+      ia_collection_s: "community;printdisabled",
+      cover_edition_key: "OL36647151M",
+      cover_i: 12539702,
+      language: ["eng", "fre"],
+      author_key: ["OL7422948A"],
+      author_name: ["James Clear"],
+      availability: {
+        status: "private",
+        available_to_browse: false,
+        available_to_borrow: false,
+        available_to_waitlist: false,
+        is_printdisabled: true,
+        is_readable: false,
+        is_lendable: false,
+        is_previewable: true,
+        identifier:
+          "atomic-habits-tiny-changes-remarkable-results-by-james-clear-z-lib.org",
+        isbn: null,
+        oclc: null,
+        openlibrary_work: "OL17930368W",
+        openlibrary_edition: "OL32336498M",
+        last_loan_date: null,
+        num_waitlist: null,
+        last_waitlist_date: null,
+        is_restricted: true,
+        is_browseable: false,
+        __src__: "core.models.lending.get_availability",
+      },
+    },
+    {
+      key: "/works/OL18020194W",
+      title: "It Ends With Us",
+      edition_count: 23,
+      first_publish_year: 2016,
+      has_fulltext: true,
+      public_scan_b: false,
+      ia: ["itendswithus0000hoov"],
+      ia_collection_s:
+        "inlibrary;internetarchivebooks;openlibrary-d-ol;printdisabled",
+      lending_edition_s: "OL27213498M",
+      lending_identifier_s: "itendswithus0000hoov",
+      cover_edition_key: "OL30570882M",
+      cover_i: 10473609,
+      language: ["eng", "ger", "pol", "por", "dut", "fre", "tur", "spa", "kor"],
+      author_key: ["OL7315784A"],
+      author_name: ["Colleen Hoover"],
+      availability: {
+        status: "borrow_unavailable",
+        available_to_browse: false,
+        available_to_borrow: false,
+        available_to_waitlist: true,
+        is_printdisabled: true,
+        is_readable: false,
+        is_lendable: true,
+        is_previewable: true,
+        identifier: "itendswithus0000hoov",
+        isbn: "9781471158254",
+        oclc: null,
+        openlibrary_work: "OL20033455W",
+        openlibrary_edition: "OL27213498M",
+        last_loan_date: null,
+        num_waitlist: null,
+        last_waitlist_date: null,
+        is_restricted: true,
+        is_browseable: false,
+        __src__: "core.models.lending.get_availability",
+      },
+    },
+    {
+      key: "/works/OL27733867W",
+      title: "It Starts with Us",
+      edition_count: 5,
+      first_publish_year: 2022,
+      has_fulltext: false,
+      public_scan_b: false,
+      cover_edition_key: "OL37857529M",
+      cover_i: 12749873,
+      language: ["eng"],
+      author_key: ["OL7315784A"],
+      author_name: ["Colleen Hoover"],
+    },
+    {
+      key: "/works/OL2010879W",
+      title: "Rich Dad, Poor Dad",
+      edition_count: 63,
+      first_publish_year: 1990,
+      has_fulltext: true,
+      public_scan_b: false,
+      ia: [
+        "richdadpoordadwh0000kiyo_c5p7",
+        "richdadpoordadwh00kiyo_0",
+        "kanemochitsanbin0000kiyo",
+        "richdadpoordadwh00kiyo",
+        "isbn_9780453677455",
+        "fubabaqiongbabar0004kiyo",
+        "paparicopapapobr0000kiyo",
+        "fubabaqiongbabar0000kiyo_g2v1",
+        "padrericopadrepo0000kiyo",
+        "richdadpoordadwh0000kiyo_t8n6",
+        "richdadpoordadwh00kiyo_843",
+      ],
+      ia_collection_s:
+        "americana;china;cnusd-ol;delawarecountydistrictlibrary;delawarecountydistrictlibrary-ol;denverpubliclibrary-ol;drakeuniversity-ol;inlibrary;internetarchivebooks;library_of_atlantis;librarygenesis;miltonpubliclibrary-ol;popularchinesebooks;printdisabled;riceuniversity-ol;rochester-ol;worthingtonlibraries-ol",
+      lending_edition_s: "OL16980586M",
+      lending_identifier_s: "richdadpoordadwh0000kiyo_c5p7",
+      cover_edition_key: "OL23243731M",
+      cover_i: 8315302,
+      language: ["eng", "jpn", "chi", "spa", "rus", "por"],
+      author_key: ["OL242220A", "OL2711033A", "OL7854295A"],
+      author_name: ["Robert T. Kiyosaki", "Sharon L. Lechter", "Tim Wheeler"],
+      availability: {
+        status: "borrow_available",
+        available_to_browse: true,
+        available_to_borrow: false,
+        available_to_waitlist: true,
+        is_printdisabled: true,
+        is_readable: false,
+        is_lendable: true,
+        is_previewable: true,
+        identifier: "richdadpoordadwh0000kiyo_c5p7",
+        isbn: "0964385619",
+        oclc: null,
+        openlibrary_work: "OL2010879W",
+        openlibrary_edition: "OL16980586M",
+        last_loan_date: null,
+        num_waitlist: null,
+        last_waitlist_date: null,
+        is_restricted: true,
+        is_browseable: true,
+        __src__: "core.models.lending.get_availability",
+      },
+    },
+  ],
+};
+
+export default trendingBooks;
