--- src/mockCoverImages.js
@@ -1,11 +1,0 @@
-const mockCoverImage1 = 'https://covers.openlibrary.org/b/id/8231990-M.jpg'
-
-const mockCoverImage2 = 'https://covers.openlibrary.org/b/id/12539702-M.jpg'
-
-const mockCoverImage3 = 'https://covers.openlibrary.org/b/id/10473609-M.jpg'
-
-const mockCoverImage4 = 'https://covers.openlibrary.org/b/id/12749873-M.jpg'
-
-const mockCoverImage5 = 'https://covers.openlibrary.org/b/id/8315302-M.jpg'
-
-export { mockCoverImage1, mockCoverImage2, mockCoverImage3, mockCoverImage4, mockCoverImage5 }
