--- src/mockDescriptions.js
@@ -1,273 +1,0 @@
-const description1 = {
-    "subjects": [
-        "Self-realization",
-        "Conduct of life",
-        "Conducta de vida",
-        "nyt:advice-how-to-and-miscellaneous=2016-10-02",
-        "New York Times bestseller"
-    ],
-    "key": "/works/OL17590212W",
-    "title": "The Subtle Art of Not Giving a F*ck",
-    "authors": [
-        {
-            "author": {
-                "key": "/authors/OL2964000A"
-            },
-            "type": {
-                "key": "/type/author_role"
-            }
-        }
-    ],
-    "type": {
-        "key": "/type/work"
-    },
-    "covers": [
-        8231990,
-        7907459,
-        8846339,
-        9358591,
-        10162546,
-        8574581,
-        8814953,
-        8596650,
-        -1,
-        8577361,
-        10483024,
-        10483025,
-        10735994,
-        12248881
-    ],
-    "latest_revision": 37,
-    "revision": 37,
-    "created": {
-        "type": "/type/datetime",
-        "value": "2016-10-26T13:42:17.105614"
-    },
-    "last_modified": {
-        "type": "/type/datetime",
-        "value": "2022-10-26T07:35:54.975819"
-    }
-}
-
-const description2 = {
-    "description": {
-        "type": "/type/text",
-        "value": "No matter your goals, Atomic Habits offers a proven framework for improving every day. James Clear, one of the world's leading experts on habit formation, reveals practical strategies that will teach you exactly how to form good habits, break bad ones, and master the tiny behaviors that lead to remarkable results."
-    },
-    "subjects": [
-        "Habit",
-        "Habit breaking",
-        "Behavior modification",
-        "Self-actualization (psychology)",
-        "Business",
-        "psychology",
-        "Personal Growth",
-        "nyt:business-books=2022-08-14",
-        "New York Times bestseller"
-    ],
-    "key": "/works/OL17930368W",
-    "title": "Atomic Habits",
-    "authors": [
-        {
-            "author": {
-                "key": "/authors/OL7422948A"
-            },
-            "type": {
-                "key": "/type/author_role"
-            }
-        }
-    ],
-    "type": {
-        "key": "/type/work"
-    },
-    "covers": [
-        12539702,
-        10457042,
-        10958382,
-        8260535,
-        12225471,
-        10510589,
-        11646973
-    ],
-    "latest_revision": 22,
-    "revision": 22,
-    "created": {
-        "type": "/type/datetime",
-        "value": "2018-10-20T04:53:43.040839"
-    },
-    "last_modified": {
-        "type": "/type/datetime",
-        "value": "2022-10-17T14:25:58.570819"
-    }
-}
-
-const description3 = {
-    "title": "It Ends With Us",
-    "key": "/works/OL18020194W",
-    "authors": [
-        {
-            "author": {
-                "key": "/authors/OL7315784A"
-            },
-            "type": {
-                "key": "/type/author_role"
-            }
-        }
-    ],
-    "type": {
-        "key": "/type/work"
-    },
-    "description": {
-        "type": "/type/text",
-        "value": "Lily hasn’t always had it easy, but that’s never stopped her from working hard for the life she wants. She’s come a long way from the small town where she grew up—she graduated from college, moved to Boston, and started her own business. And when she feels a spark with a gorgeous neurosurgeon named Ryle Kincaid, everything in Lily’s life seems too good to be true.\r\n\r\nRyle is assertive, stubborn, maybe even a little arrogant. He’s also sensitive, brilliant, and has a total soft spot for Lily. And the way he looks in scrubs certainly doesn’t hurt. Lily can’t get him out of her head. But Ryle’s complete aversion to relationships is disturbing. Even as Lily finds herself becoming the exception to his “no dating” rule, she can’t help but wonder what made him that way in the first place.\r\n\r\nAs questions about her new relationship overwhelm her, so do thoughts of Atlas Corrigan—her first love and a link to the past she left behind. He was her kindred spirit, her protector. When Atlas suddenly reappears, everything Lily has built with Ryle is threatened.\r\n\r\nWith this bold and deeply personal novel, It Ends With Us is a heart-wrenching story and an unforgettable tale of love that comes at the ultimate price."
-    },
-    "covers": [
-        10473609,
-        12501970,
-        10126667,
-        12498414,
-        12503959,
-        12209375
-    ],
-    "subject_places": [
-        "Boston (Mass.)"
-    ],
-    "subjects": [
-        "Fiction, Romance, Contemporary, Life change events",
-        "Triangles (Interpersonal relations)",
-        "Man-woman relationships",
-        "Neurosurgeons",
-        "First loves",
-        "Businesswomen",
-        "Fiction",
-        "Contemporary Women",
-        "New York Times bestseller",
-        "Fiction, romance, contemporary",
-        "Man-woman relationships, fiction",
-        "Fiction, women",
-        "Boston (mass.), fiction"
-    ],
-    "latest_revision": 27,
-    "revision": 27,
-    "created": {
-        "type": "/type/datetime",
-        "value": "2018-12-06T15:20:02.050673"
-    },
-    "last_modified": {
-        "type": "/type/datetime",
-        "value": "2022-10-26T07:09:51.468001"
-    }
-}
-
-const description4 = {
-    "title": "It Starts with Us",
-    "authors": [
-        {
-            "author": {
-                "key": "/authors/OL7315784A"
-            },
-            "type": {
-                "key": "/type/author_role"
-            }
-        }
-    ],
-    "key": "/works/OL27733867W",
-    "type": {
-        "key": "/type/work"
-    },
-    "covers": [
-        12749873
-    ],
-    "subjects": [
-        "Contemporary Romance",
-        "Friendship, fiction"
-    ],
-    "latest_revision": 5,
-    "revision": 5,
-    "created": {
-        "type": "/type/datetime",
-        "value": "2022-05-06T15:17:33.922619"
-    },
-    "last_modified": {
-        "type": "/type/datetime",
-        "value": "2022-10-10T18:20:18.500882"
-    }
-}
-
-const description5 = {
-    "description": {
-        "type": "/type/text",
-        "value": "Argues that a good education and a secure job are not guarantees for financial success, and describes six guidelines for making money work for oneself."
-    },
-    "title": "Rich Dad, Poor Dad",
-    "covers": [
-        8315302,
-        8715673,
-        -1,
-        10483841,
-        10537463,
-        12506734
-    ],
-    "first_publish_date": "January 3, 2002",
-    "key": "/works/OL2010879W",
-    "authors": [
-        {
-            "author": {
-                "key": "/authors/OL242220A"
-            },
-            "type": {
-                "key": "/type/author_role"
-            }
-        },
-        {
-            "author": {
-                "key": "/authors/OL2711033A"
-            },
-            "type": {
-                "key": "/type/author_role"
-            }
-        },
-        {
-            "author": {
-                "key": "/authors/OL7854295A"
-            },
-            "type": {
-                "key": "/type/author_role"
-            }
-        }
-    ],
-    "type": {
-        "key": "/type/work"
-    },
-    "subjects": [
-        "Rich people",
-        "Personal Finance",
-        "Investments",
-        "Finanzas personales",
-        "Inversiones",
-        "Ricos",
-        "Riqueza",
-        "Financial security",
-        "Success",
-        "Financiering",
-        "Huishoudingen",
-        "Finance, personal",
-        "Hacienda privada",
-        "Spanish language materials"
-    ],
-    "dewey_number": [
-        "332.024"
-    ],
-    "latest_revision": 21,
-    "revision": 21,
-    "created": {
-        "type": "/type/datetime",
-        "value": "2009-12-09T22:42:37.456780"
-    },
-    "last_modified": {
-        "type": "/type/datetime",
-        "value": "2022-10-18T15:25:28.947307"
-    }
-}
-
-export { description1, description2, description3, description4, description5 }
